+++ rswag-specs/spec/rswag/specs/response_validator_spec.rb
@@ -8,15 +8,15 @@ module Rswag
       subject { ResponseValidator.new(config) }
 
       before do
-        allow(config).to receive(:get_swagger_doc).and_return(swagger_doc)
-        allow(config).to receive(:get_swagger_doc_version).and_return('2.0')
-        allow(config).to receive(:swagger_strict_schema_validation).and_return(swagger_strict_schema_validation)
+        allow(config).to receive(:get_openapi_spec).and_return(openapi_spec)
+        allow(config).to receive(:get_openapi_spec_version).and_return('2.0')
+        allow(config).to receive(:openapi_strict_schema_validation).and_return(openapi_strict_schema_validation)
       end
 
       let(:config) { double('config') }
-      let(:swagger_doc) { {} }
+      let(:openapi_spec) { {} }
       let(:example) { double('example') }
-      let(:swagger_strict_schema_validation) { false }
+      let(:openapi_strict_schema_validation) { false }
       let(:metadata) do
         {
           response: {
@@ -53,7 +53,7 @@ module Rswag
         let(:response) do
           OpenStruct.new(
             code: '200',
-            headers: { 
+            headers: {
               'X-Rate-Limit-Limit' => '10',
               'X-Cursor' => 'test_cursor',
               'X-Per-Page' => 25
@@ -77,8 +77,8 @@ module Rswag
         end
 
         context 'response headers do not include optional header' do
-          before { 
-            response.headers = { 
+          before {
+            response.headers = {
               'X-Rate-Limit-Limit' => '10',
               'X-Per-Page' => 25
             }
@@ -87,8 +87,8 @@ module Rswag
         end
 
         context 'response headers include nullable header' do
-          before { 
-            response.headers = { 
+          before {
+            response.headers = {
               'X-Rate-Limit-Limit' => '10',
               'X-Cursor' => 'test_cursor',
               'X-Per-Page' => nil
@@ -98,8 +98,8 @@ module Rswag
         end
 
         context 'response headers missing nullable header' do
-          before { 
-            response.headers = { 
+          before {
+            response.headers = {
               'X-Rate-Limit-Limit' => '10',
               'X-Cursor' => 'test_cursor'
             }
@@ -116,27 +116,27 @@ module Rswag
           before { response.body = '{"foo":"Some comment", "number": 3, "text":"bar"}' }
 
           context "with strict schema validation enabled" do
-            let(:swagger_strict_schema_validation) { true }
+            let(:openapi_strict_schema_validation) { true }
 
             it { expect { call }.to raise_error /Expected response body/ }
           end
 
           context "with strict schema validation disabled" do
-            let(:swagger_strict_schema_validation) { false }
+            let(:openapi_strict_schema_validation) { false }
 
             it { expect { call }.not_to raise_error }
           end
 
           context "with strict schema validation disabled in config but enabled in metadata" do
-            let(:swagger_strict_schema_validation) { false }
-            let(:metadata) { super().merge(swagger_strict_schema_validation: true) }
+            let(:openapi_strict_schema_validation) { false }
+            let(:metadata) { super().merge(openapi_strict_schema_validation: true) }
 
             it { expect { call }.to raise_error /Expected response body/ }
           end
 
           context "with strict schema validation enabled in config but disabled in metadata" do
-            let(:swagger_strict_schema_validation) { true }
-            let(:metadata) { super().merge(swagger_strict_schema_validation: false) }
+            let(:openapi_strict_schema_validation) { true }
+            let(:metadata) { super().merge(openapi_strict_schema_validation: false) }
 
             it { expect { call }.not_to raise_error }
           end
@@ -145,7 +145,7 @@ module Rswag
         context 'referenced schemas' do
           context 'swagger 2.0' do
             before do
-              swagger_doc[:definitions] = {
+              openapi_spec[:definitions] = {
                 'blog' => {
                   type: :object,
                   properties: { foo: { type: :string } },
@@ -163,9 +163,9 @@ module Rswag
           context 'openapi 3.0.1' do
             context 'components/schemas' do
               before do
-                allow(ActiveSupport::Deprecation).to receive(:warn)
-                allow(config).to receive(:get_swagger_doc_version).and_return('3.0.1')
-                swagger_doc[:components] = {
+                allow(Rswag::Specs.deprecator).to receive(:warn)
+                allow(config).to receive(:get_openapi_spec_version).and_return('3.0.1')
+                openapi_spec[:components] = {
                   schemas: {
                     'blog' => {
                       type: :object,
@@ -185,7 +185,7 @@ module Rswag
                 let(:response) do
                   OpenStruct.new(
                     code: '200',
-                    headers: { 
+                    headers: {
                       'X-Rate-Limit-Limit' => '10',
                       'X-Cursor' => 'test_cursor',
                       'X-Per-Page' => 25
@@ -226,7 +226,7 @@ module Rswag
                 let(:response) do
                   OpenStruct.new(
                     code: '200',
-                    headers: { 
+                    headers: {
                       'X-Rate-Limit-Limit' => '10',
                       'X-Cursor' => 'test_cursor',
                       'X-Per-Page' => 25
@@ -270,9 +270,9 @@ module Rswag
 
             context 'deprecated definitions' do
               before do
-                allow(ActiveSupport::Deprecation).to receive(:warn)
-                allow(config).to receive(:get_swagger_doc_version).and_return('3.0.1')
-                swagger_doc[:definitions] = {
+                allow(Rswag::Specs.deprecator).to receive(:warn)
+                allow(config).to receive(:get_openapi_spec_version).and_return('3.0.1')
+                openapi_spec[:definitions] = {
                   'blog' => {
                     type: :object,
                     properties: { foo: { type: :string } },
@@ -284,7 +284,7 @@ module Rswag
 
               it 'warns the user to upgrade' do
                 expect { call }.to raise_error(/Expected response body/)
-                expect(ActiveSupport::Deprecation).to have_received(:warn)
+                expect(Rswag::Specs.deprecator).to have_received(:warn)
                   .with('Rswag::Specs: WARNING: definitions is replaced in OpenAPI3! Rename to components/schemas (in swagger_helper.rb)')
               end
             end