+++ test-app/spec/integration/blogs_spec.rb
@@ -1,10 +1,10 @@
 require 'swagger_helper'
 
-RSpec.describe 'Blogs API', type: :request, swagger_doc: 'v1/swagger.json' do
+RSpec.describe 'Blogs API', type: :request, openapi_spec: 'v1/swagger.json' do
   let(:api_key) { 'fake_key' }
 
   before do
-    allow(ActiveSupport::Deprecation).to receive(:warn) # Silence deprecation output from specs
+    # allow(Rswag::Specs.deprecator).to receive(:warn) # Silence deprecation output from specs
   end
 
   path '/blogs' do
@@ -14,7 +14,7 @@ RSpec.describe 'Blogs API', type: :request, swagger_doc: 'v1/swagger.json' do
       operationId 'createBlog'
       consumes 'application/json'
       produces 'application/json'
-      parameter name: :blog, in: :body, schema: { '$ref' => '#/definitions/blog' }
+      parameter name: :blog, in: :body, schema: { '$ref' => '#/components/schemas/blog' }
 
       let(:blog) { { title: 'foo', content: 'bar', status: 'published' } }
 
@@ -24,7 +24,7 @@ RSpec.describe 'Blogs API', type: :request, swagger_doc: 'v1/swagger.json' do
       end
 
       response "422", "invalid request" do
-        schema "$ref" => "#/definitions/errors_object"
+        schema "$ref" => "#/components/schemas/errors_object"
 
         let(:blog) { {title: "foo"} }
 
@@ -55,7 +55,7 @@ RSpec.describe 'Blogs API', type: :request, swagger_doc: 'v1/swagger.json' do
       let(:blog_status) { 'published' }
 
       response '200', 'success' do
-        schema type: 'array', items: { '$ref' => '#/definitions/blog' }
+        schema type: 'array', items: { '$ref' => '#/components/schemas/blog' }
 
         run_test! do
           expect(JSON.parse(response.body).size).to eq(1)
@@ -63,7 +63,7 @@ RSpec.describe 'Blogs API', type: :request, swagger_doc: 'v1/swagger.json' do
       end
 
       response '200', 'no content' do
-        schema type: 'array', items: { '$ref' => '#/definitions/blog' }
+        schema type: 'array', items: { '$ref' => '#/components/schemas/blog' }
 
         let(:blog_status) { 'invalid' }
 
@@ -97,7 +97,7 @@ RSpec.describe 'Blogs API', type: :request, swagger_doc: 'v1/swagger.json' do
       let(:flexible_blog) { { blog: { headline: 'my headline', text: 'my text' } } }
 
       response '201', 'flexible blog created' do
-        schema oneOf: [{ '$ref' => '#/definitions/blog' }, { '$ref' => '#/definitions/flexible_blog' }]
+        schema oneOf: [{ '$ref' => '#/components/schemas/blog' }, { '$ref' => '#/components/schemas/flexible_blog' }]
         run_test!
       end
     end
@@ -120,7 +120,7 @@ RSpec.describe 'Blogs API', type: :request, swagger_doc: 'v1/swagger.json' do
         header 'Last-Modified', type: :string
         header 'Cache-Control', type: :string
 
-        schema '$ref' => '#/definitions/blog'
+        schema '$ref' => '#/components/schemas/blog'
 
         #Legacy
         examples 'application/json' => {
@@ -148,8 +148,8 @@ RSpec.describe 'Blogs API', type: :request, swagger_doc: 'v1/swagger.json' do
 
         run_test!
 
-        context 'when swagger_strict_schema_validation is true' do
-          run_test!(swagger_strict_schema_validation: true)
+        context 'when openapi_strict_schema_validation is true' do
+          run_test!(openapi_strict_schema_validation: true)
         end
       end
 
@@ -158,8 +158,8 @@ RSpec.describe 'Blogs API', type: :request, swagger_doc: 'v1/swagger.json' do
         run_test!
       end
 
-      response '200', 'blog found - swagger_strict_schema_validation = true', swagger_strict_schema_validation: true do
-        schema '$ref' => '#/definitions/blog'
+      response '200', 'blog found - openapi_strict_schema_validation = true', openapi_strict_schema_validation: true do
+        schema '$ref' => '#/components/schemas/blog'
 
         let(:id) { blog.id }